Base of websocket setted
