+++ api/package.json
@@ -7,8 +7,9 @@
         "cors": "^2.8.5",
         "dotenv": "^16.4.5",
         "express": "^4.21.1",
+        "express-ws": "^5.0.2",
         "pg": "^8.13.0",
-        "socket.io-client": "^4.8.1",
+        "socket.io": "^4.8.1",
         "uninstall": "^0.0.0",
         "your-node-app-name": "file:",
         "zod": "^3.23.8"
@@ -16,6 +17,7 @@
     "devDependencies": {
         "@types/cors": "^2.8.17",
         "@types/express": "^5.0.0",
+        "@types/express-ws": "^3.0.5",
         "@types/node": "latest",
         "@types/pg": "^8.11.10",
         "nodemon": "^3.1.7",
