+++ api/src/http/models/user.model.ts
@@ -1,6 +1,6 @@
 import { IUser } from "../interfaces/interfaces";
 
-import { dbConnect } from '../conf/db';
+import { dbConnect } from '../../conf/db';
 
 export class UserModel {
